added custom errors and error handler
