+++ errors.js
@@ -1,0 +1,13 @@
+
+
+class InvalidConfigError extends Error {
+    constructor(message) {
+        super(message);
+        this.name = 'InvalidConfigError';
+        this.isCustom = true;
+    }
+}
+
+
+
+export { InvalidConfigError };
